+++ javafxFactoryEditing/src/test/java/io/github/factoryfx/javafx/editor/data/Main.java
@@ -1,0 +1,8 @@
+package io.github.factoryfx.javafx.editor.data;
+
+public class Main {
+
+    public static void main(String[] args) {
+        DataEditorIntegrationTest.main(args);
+    }
+}