+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+# 4.1.7
+* **Update dependencies**
+* **Remove copper bridge**
+* **ValueListAttributeVisualisation**
+    * minor improvements
+
 # 4.1.6
 * **ValueListAttributeVisualisation**
-    * fix readOnly not being applied to value list attribute visualisation
+    * fix readOnly not being applied to value list attribute visualization
 
 # 4.1.5
 * **DataStorage**